+++ packages/k2-alpine/src/components/Chart/PriceChart/helpers.ts
@@ -110,7 +110,7 @@ export const traceSmoothLine = (
   }
 }
 
-export type CandleDisplayStrings = {
+type CandleDisplayStrings = {
   priceText: string
   timeText: string
   deltaPriceText: string
